--- src/ConsoleApp/bin/Debug/net5.0/ConsoleApp.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v5.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v5.0": {
-      "ConsoleApp/1.0.0": {
-        "runtime": {
-          "ConsoleApp.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "ConsoleApp/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
