--- src/ConsoleApp/bin/Debug/net5.0/ConsoleApp.runtimeconfig.dev.json
@@ -1,8 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\flpin\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\flpin\\.nuget\\packages"
-    ]
-  }
-}
