--- src/ConsoleApp/bin/Debug/net5.0/ConsoleApp.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "net5.0",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "5.0.0"
-    }
-  }
-}
